Updated cli.
