+++ cli/.buukrc.json
@@ -1,3 +1,3 @@
 {
-  "base_path": "/Users/Hang/Desktop/test2"
+  "base_path": "/Users/Hang/Desktop/test"
 }
