+++ cli/manifest.json
@@ -1,5 +1,5 @@
 {
-  "name": "Buuk",
+  "name": "Test",
   "short_name": "Buuk",
   "description": "A Buuk project.",
   "articles": {}
